requirements.txt
# native python support
# subprocess
# os
# time
# gzip
# shutil
python-dotenv